Avoid using alarm(2) for timeouts
libvchan has much better ways of dealing with timeouts now, so use them
instead of calling an async signal unsafe function from a signal
handler.  Both qrexec-agent-data.c and qrexec-daemon-common.c
reimplemented timeouts, so write common code in libqrexec for both to
use.
